+++ libs/wallet-react/src/components/connect-dialog/connect-dialog/connect-dialog.tsx
@@ -7,7 +7,7 @@ import { useWallet } from '../../../hooks/use-wallet';
 import { useConnect } from '../../../hooks/use-connect';
 import { RiskAck } from '../../risk-ack';
 import { ConnectionStatus } from '../connection-status';
-import { ConnectionOptions } from '../connection-options';
+import { ConnectionOptionsList } from '../connection-options';
 
 export const DIALOG_CLOSE_DELAY = 1000;
 
@@ -74,7 +74,7 @@ const Content = ({
   }
 
   if (status === 'disconnected') {
-    return <ConnectionOptions onConnect={onConnect} />;
+    return <ConnectionOptionsList onConnect={onConnect} />;
   }
 
   return <ConnectionStatus status={status} />;
@@ -100,7 +100,7 @@ export const ConnectDialog = ({
   return (
     <Dialog open={open} size="small" onChange={onChange}>
       {status === 'disconnected' ? (
-        <ConnectionOptions onConnect={onConnect} />
+        <ConnectionOptionsList onConnect={onConnect} />
       ) : (
         <ConnectionStatus status={status} />
       )}
