+++ libs/wallet-react/src/components/connect-dialog/connection-options/connection-options-list/prominent-connectors-list.tsx
@@ -1,6 +1,6 @@
 import { type Connector, type ConnectorType } from '@vegaprotocol/wallet';
-import { ConnectionOptionRecord } from './custom-connector-options';
-import { ConnectionOptionDefault } from './connection-option-default';
+import { ConnectionOptionRecord } from '../custom-connector-options';
+import { ConnectionOptionDefault } from '../connection-option-default';
 
 export const ProminentConnectorsList = ({
   onConnect,
