+++ libs/wallet-react/src/components/connect-dialog/connection-options/custom-connector-options/index.tsx
@@ -1,17 +1,8 @@
 import { type FunctionComponent } from 'react';
-import { type Connector, type ConnectorType } from '@vegaprotocol/wallet';
+import { type ConnectorType } from '@vegaprotocol/wallet';
 import { ConnectionOptionInjected } from './connection-option-injected';
 import { ConnectionOptionQuickstart } from './connection-option-quickstart';
-
-interface ConnectionOptionProps {
-  id: ConnectorType;
-  name: string;
-  description: string;
-  showDescription?: boolean;
-  onClick: () => void;
-  onInstall?: () => void;
-  connector: Connector;
-}
+import { type ConnectionOptionProps } from '../types';
 
 export const ConnectionOptionRecord: {
   [C in ConnectorType]?: FunctionComponent<ConnectionOptionProps>;
