+++ libs/wallet-react/src/components/connect-dialog/connection-options/types.ts
@@ -1,0 +1,11 @@
+import { type Connector, type ConnectorType } from '@vegaprotocol/wallet';
+
+export interface ConnectionOptionProps {
+  id: ConnectorType;
+  name: string;
+  description: string;
+  showDescription?: boolean;
+  onClick: () => void;
+  onInstall?: () => void;
+  connector: Connector;
+}
