+++ libs/wallet-react/src/components/connect-dialog/index.ts
@@ -1,5 +1,5 @@
 export { ConnectDialogWithRiskAck, ConnectDialog } from './connect-dialog';
-export { ConnectionOptions } from './connection-options';
+export { ConnectionOptionsList as ConnectionOptions } from './connection-options';
 export { ConnectionStatus } from './connection-status';
 export { ConnectorIcon } from './connection-options/connector-icon';
 export { ConnectionOptionDefault } from './connection-options';
